Rename `stop_tactic.cpp` and `stop_tactic.h` to `halt_tactic.cpp` and `halt_tactic.h`, renaming variables and strings where necessary
Variables and strings renamed for consistency and to allow tests to successfully run
