+++ src/software/ai/hl/stp/play/test_plays/BUILD
@@ -6,7 +6,7 @@ cc_library(
     hdrs = ["halt_test_play.h"],
     deps = [
         "//software/ai/hl/stp/play",
-        "//software/ai/hl/stp/tactic/stop:stop_tactic",
+        "//software/ai/hl/stp/tactic/stop:halt_tactic",
         "//software/util/generic_factory",
     ],
     alwayslink = True,
