+++ src/software/ai/hl/stp/play/test_plays/halt_test_play.cpp
@@ -1,6 +1,6 @@
 #include "software/ai/hl/stp/play/test_plays/halt_test_play.h"
 
-#include "software/ai/hl/stp/tactic/stop/stop_tactic.h"
+#include "software/ai/hl/stp/tactic/stop/halt_tactic.h"
 #include "software/geom/algorithms/contains.h"
 #include "software/util/generic_factory/generic_factory.h"
