+++ src/software/ai/hl/stp/tactic/stop/BUILD
@@ -1,14 +1,14 @@
 package(default_visibility = ["//visibility:public"])
 
 cc_library(
-    name = "stop_tactic",
+    name = "halt_tactic",
     srcs = [
         "stop_fsm.cpp",
-        "stop_tactic.cpp",
+        "halt_tactic.cpp",
     ],
     hdrs = [
         "stop_fsm.h",
-        "stop_tactic.h",
+        "halt_tactic.h",
     ],
     deps = [
         "//shared:constants",
@@ -21,7 +21,7 @@ cc_test(
     name = "stop_fsm_test",
     srcs = ["stop_fsm_test.cpp"],
     deps = [
-        ":stop_tactic",
+        ":halt_tactic",
         "//shared/test_util:tbots_gtest_main",
         "//software/test_util",
     ],
@@ -31,7 +31,7 @@ cc_test(
     name = "stop_tactic_test",
     srcs = ["stop_tactic_test.cpp"],
     deps = [
-        ":stop_tactic",
+        ":halt_tactic",
         "//shared/test_util:tbots_gtest_main",
         "//software/simulated_tests:simulated_er_force_sim_play_test_fixture",
         "//software/simulated_tests/terminating_validation_functions",
