+++ src/software/ai/hl/stp/tactic/stop/halt_tactic.h
