+++ src/software/field_tests/passing_field_test.py
@@ -96,11 +96,11 @@ def test_passing(field_test_runner):
         test_timeout_s=5,
     )
 
-    # Send a stop tactic after the test finishes
-    stop_tactic = StopTactic()
+    # Send a halt tactic after the test finishes
+    halt_tactic = StopTactic()
     params = AssignedTacticPlayControlParams()
-    params.assigned_tactics[passer_robot_id].stop.CopyFrom(stop_tactic)
-    params.assigned_tactics[receiver_robot_id].stop.CopyFrom(stop_tactic)
+    params.assigned_tactics[passer_robot_id].stop.CopyFrom(halt_tactic)
+    params.assigned_tactics[receiver_robot_id].stop.CopyFrom(halt_tactic)
     # send the stop tactic
     field_test_runner.set_tactics(params, True)
