+++ src/software/field_tests/pivot_kick_field_test.py
@@ -38,10 +38,10 @@ def test_pivot_kick(field_test_runner):
         eventually_validation_sequence_set=[[]],
         test_timeout_s=15,
     )
-    # Send a stop tactic after the test finishes
-    stop_tactic = StopTactic()
+    # Send a halt tactic after the test finishes
+    halt_tactic = HaltTactic()
     params = AssignedTacticPlayControlParams()
-    params.assigned_tactics[id].stop.CopyFrom(stop_tactic)
+    params.assigned_tactics[id].stop.CopyFrom(halt_tactic)
 
 
 if __name__ == "__main__":
